Ignore wipe operation on non-block devices.
The implementation of make_ext4fs() calls make_ext4fs_internal() by
forcing the |wipe| parameter to true, which is problematic when the
library is used within the emulator (the wipe operation will always
fail on non-Linux platforms).

The wipe is only useful when writing directly to a block device,
so only call wipe_block_device() when it is certain that the file
descriptor points to a block device.

Also convert the fatal error into a warning, when trying to actually
wipe a block device on non-Linux systems.

BUG=NONE

+++ ext4_utils/make_ext4fs.c
@@ -623,8 +623,13 @@ int make_ext4fs_internal(int fd, const char *_directory,
 			aux_info.sb->s_blocks_count_lo - aux_info.sb->s_free_blocks_count_lo,
 			aux_info.sb->s_blocks_count_lo);
 
-	if (wipe)
-		wipe_block_device(fd, info.len);
+	if (wipe) {
+		if (!is_block_device_fd(fd)) {
+			printf("Ignoring wipe on non-block device file\n");
+		} else {
+			wipe_block_device(fd, info.len);
+		}
+	}
 
 	write_ext4_image(fd, gzip, sparse, crc);
